nexuszero-optimizer/requirements.txt
# Core ML
torch>=2.1.0
torch-geometric>=2.4.0
torch-scatter>=2.1.2
torch-sparse>=0.6.18

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
networkx>=3.1

# Optimization
optuna>=3.3.0
ray[tune]>=2.7.0

# Rust Bridge (optional - for production)
# maturin>=1.3.0
# pyo3>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
h5py>=3.9.0

# Development
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0
ruff>=0.0.285
